First commit of Lucas Number generator.
This project generates Lucas numbers. The sequence is limited by the
maximum Lucas number required (threshold reached). The golden ratio is
also calculated for good measure however not tested for in the unit
test as a correct Lucas sequence by default computes to the correct
ratio. The unit test only looks to the first 10 enties for validation.