Made changes as per PR with JSmith.
__init__.py : Added import for golden_ratio_calc
test_lucas_number.py : Chnaged assert for unit test when verifying
lucas sequence.

Added signal averager using decorators. These are separate files rather than an overwrite so that a comparison can be made easily. The intention is not to merge into one file but keep as separate examples.

Added fix to SigAverager using decorators

This commit added a fix to the recursion of the return value used when
doing signal averaging.

No SARAO JIRA ticket as it's a private repo.

Added a realisation to the README. It should only to this branch.